impermax: refactor shared code to helper folder
