+++ src/apps/impermax/polygon/impermax.lend.token-fetcher.ts
@@ -1,13 +1,11 @@
 import { Inject } from '@nestjs/common';
 
-import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { Register } from '~app-toolkit/decorators';
 import { PositionFetcher } from '~position/position-fetcher.interface';
 import { AppTokenPosition } from '~position/position.interface';
 import { Network } from '~types/network.interface';
 
-import { ImpermaxContractFactory, Borrowable } from '../contracts';
-import { getBorrowAddresses } from '../ethereum/impermax.lend.token-fetcher';
+import { ImpermaxLendTokenHelper } from '../helpers/impermax.lend.token-fetcher-helper';
 import { IMPERMAX_DEFINITION } from '../impermax.definition';
 
 const appId = IMPERMAX_DEFINITION.id;
@@ -17,30 +15,9 @@ export const address = '0xBB92270716C8c424849F17cCc12F4F24AD4064D6'.toLowerCase(
 
 @Register.TokenPositionFetcher({ appId, groupId, network })
 export class PolygonImpermaxLendTokenFetcher implements PositionFetcher<AppTokenPosition> {
-  constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(ImpermaxContractFactory) private readonly contractFactory: ImpermaxContractFactory,
-  ) {}
+  constructor(@Inject(ImpermaxLendTokenHelper) private readonly impermaxLendTokenHelper: ImpermaxLendTokenHelper) {}
 
   async getPositions() {
-    const contract = this.contractFactory.factory({ address, network });
-    const multicall = this.appToolkit.getMulticall(network).wrap(contract);
-    const collateralAddresses = await getBorrowAddresses(multicall);
-
-    const tokens = await this.appToolkit.helpers.vaultTokenHelper.getTokens<Borrowable>({
-      appId,
-      groupId,
-      network,
-      resolveVaultAddresses: () => collateralAddresses,
-      resolveContract: ({ address, network }) => this.contractFactory.borrowable({ address, network }),
-      resolveUnderlyingTokenAddress: ({ multicall, contract }) => multicall.wrap(contract).underlying(),
-      resolveReserve: () => 0,
-      resolvePricePerShare: ({ multicall, contract, underlyingToken }) =>
-        multicall
-          .wrap(contract)
-          .exchangeRateLast()
-          .then(rate => Number(rate) / 10 ** underlyingToken.decimals),
-    });
-    return tokens;
+    return this.impermaxLendTokenHelper.getPositions({ address, network });
   }
 }
